+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+[comment]: <> (## [Unreleased])
+## [3.2.0] - 2021-04-11
 ### BREAKING CHANGES
 - sortArrayByObjectProperty no longer accepts 'undefined' as sortable attribute.
 - Removed deprecated functionality from objectsArrayToObject.
@@ -60,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial code.
 - Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/array-utils/compare/3.2.0...HEAD
+[Unreleased]: https://github.com/Ionaru/array-utils/compare/4.0.0...HEAD
+[4.0.0]: https://github.com/Ionaru/array-utils/compare/3.2.0...4.0.0
 [3.2.0]: https://github.com/Ionaru/array-utils/compare/3.1.0...3.2.0
 [3.1.0]: https://github.com/Ionaru/array-utils/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/Ionaru/array-utils/compare/2.0.0...3.0.0